more tests, better sqlalchemy2 and pydantic2 code, some refactor

+++ services/my_model.py
@@ -1,63 +1,77 @@
 from datetime import datetime
+from typing import Optional
 
-from sqlalchemy.orm import Session
-from sqlalchemy.sql.expression import func
+from sqlalchemy import select
+from sqlalchemy.ext.asyncio import AsyncSession
 
 from helpers.log import logging as l
 from models.my_model import MyModel
 
 
-def create(obj: MyModel, db: Session):
+async def create(obj: MyModel, db: AsyncSession) -> Optional[int]:
     obj.created_at = datetime.now()
     try:
         db.add(obj)
-        db.commit()
-        db.refresh(obj)
+        await db.commit()
+        await db.refresh(obj)
         return obj.id
     except Exception as e:
         l.error(f"[my model : create] {e}")
-        db.rollback()
+        await db.rollback()
         return None
 
 
-def get_random_row(db: Session):
+async def get_random_row(db: AsyncSession) -> Optional[MyModel]:
     try:
-        random_row = db.query(MyModel).order_by(func.random()).first()
+        stmt = select(MyModel).order_by(MyModel.id.desc()).limit(1)
+        result = await db.execute(stmt)
+        random_row = result.scalar_one_or_none()
         return random_row
     except Exception as e:
         l.error(f"[my model : get random row] {e}")
         return None
 
 
-def update(id, obj: MyModel, db: Session):
+async def update(id: int, obj: MyModel, db: AsyncSession) -> Optional[MyModel]:
     try:
-        item = db.query(MyModel).filter(MyModel.id == id).one()
-        item.field1 = obj.field1
-        item.field2 = obj.field2
-        item.updated_at = datetime.now()
-        db.commit()
-        return item
+        stmt = select(MyModel).where(MyModel.id == id)
+        result = await db.execute(stmt)
+        item = result.scalar_one_or_none()
+        if item:
+            item.field1 = obj.field1
+            item.field2 = obj.field2
+            item.updated_at = datetime.now()
+            await db.commit()
+            await db.refresh(item)
+            return item
+        return None
     except Exception as e:
         l.error(f"[my model : update] {e}")
-        db.rollback()
-    return None
+        await db.rollback()
+        return None
 
 
-def delete(id, db: Session):
+async def delete(id: int, db: AsyncSession) -> bool:
     try:
-        item = db.query(MyModel).filter(MyModel.id == id).one()
-        db.delete(item)
-        db.commit()
-        return True
+        stmt = select(MyModel).where(MyModel.id == id)
+        result = await db.execute(stmt)
+        item = result.scalar_one_or_none()
+        if item:
+            await db.delete(item)
+            await db.commit()
+            return True
+        return False
     except Exception as e:
         l.error(f"[my model : delete] {e}")
-        db.rollback()
+        await db.rollback()
         return False
 
 
-def find_by_id(id, db: Session):
+async def find_by_id(id: int, db: AsyncSession) -> Optional[MyModel]:
     try:
-        item = db.query(MyModel).filter(MyModel.id == id).one()
+        stmt = select(MyModel).where(MyModel.id == id)
+        result = await db.execute(stmt)
+        item = result.scalar_one_or_none()
         return item
     except Exception as e:
         l.error(f"[my model : find by id] {e}")
